+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 7.7
+--------------
+- Move Dashboard() call outside in Dashboard.fnc.php & Portal.php
+
 Changes in 7.6.1
 ----------------
 - Fix #307 XSS update CodeIgniter Security class in classes/Security.php, thanks to @DustinBorn